package name change to achievement and user entity change
