--- src/main/java/com/alom/dorundorunbe/domain/Achievement/domain/RewardType.java
@@ -1,9 +1,0 @@
-package com.alom.dorundorunbe.domain.Achievement.domain;
-
-
-
-
-public enum RewardType {
-    COIN, BACKGROUND
-
-}
