--- src/main/java/com/alom/dorundorunbe/domain/Achievement/dto/reward/RewardAchievementRequestDto.java
@@ -1,10 +1,0 @@
-package com.alom.dorundorunbe.domain.Achievement.dto.reward;
-
-public record RewardAchievementRequestDto(
-        Long userId,
-        Long achievementId
-) {
-    public static RewardAchievementRequestDto of(Long userId, Long achievementId) {
-        return new RewardAchievementRequestDto(userId, achievementId);
-    }
-}
