--- src/main/java/com/alom/dorundorunbe/domain/Achievement/exception/AchievementAlreadyExistsException.java
@@ -1,8 +1,0 @@
-package com.alom.dorundorunbe.domain.Achievement.exception;
-
-
-public class AchievementAlreadyExistsException extends RuntimeException {
-    public AchievementAlreadyExistsException(String message) {
-        super(message);
-    }
-}
