--- src/main/java/com/alom/dorundorunbe/domain/Achievement/exception/AchievementConditionNotMetException.java
@@ -1,7 +1,0 @@
-package com.alom.dorundorunbe.domain.Achievement.exception;
-
-public class AchievementConditionNotMetException extends RuntimeException {
-    public AchievementConditionNotMetException(String message) {
-        super(message);
-    }
-}
