--- src/main/java/com/alom/dorundorunbe/domain/Achievement/exception/AchievementNotFoundException.java
@@ -1,7 +1,0 @@
-package com.alom.dorundorunbe.domain.Achievement.exception;
-
-public class AchievementNotFoundException extends RuntimeException {
-    public AchievementNotFoundException(String message) {
-        super(message);
-    }
-}
