--- src/main/java/com/alom/dorundorunbe/domain/Achievement/exception/RewardAlreadyClaimedException.java
@@ -1,7 +1,0 @@
-package com.alom.dorundorunbe.domain.Achievement.exception;
-
-public class RewardAlreadyClaimedException extends RuntimeException {
-    public RewardAlreadyClaimedException(String message) {
-        super(message);
-    }
-}
