--- src/main/java/com/alom/dorundorunbe/domain/Achievement/exception/UserAchievementAlreadyClaimedException.java
@@ -1,7 +1,0 @@
-package com.alom.dorundorunbe.domain.Achievement.exception;
-
-public class UserAchievementAlreadyClaimedException extends RuntimeException {
-    public UserAchievementAlreadyClaimedException(String message) {
-        super(message);
-    }
-}
