--- src/main/java/com/alom/dorundorunbe/domain/Achievement/exception/UserAchievementNotFoundException.java
@@ -1,7 +1,0 @@
-package com.alom.dorundorunbe.domain.Achievement.exception;
-
-public class UserAchievementNotFoundException extends RuntimeException {
-    public UserAchievementNotFoundException(String message) {
-        super(message);
-    }
-}
