--- src/main/java/com/alom/dorundorunbe/domain/Achievement/exception/UserNotFoundException.java
@@ -1,7 +1,0 @@
-package com.alom.dorundorunbe.domain.Achievement.exception;
-
-public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException(String message) {
-        super(message);
-    }
-}
