+++ src/main/java/com/alom/dorundorunbe/domain/achievement/domain/RewardType.java
@@ -1,0 +1,9 @@
+package com.alom.dorundorunbe.domain.achievement.domain;
+
+
+
+
+public enum RewardType {
+    COIN, BACKGROUND
+
+}
