+++ src/main/java/com/alom/dorundorunbe/domain/achievement/dto/reward/RewardAchievementRequestDto.java
@@ -1,0 +1,10 @@
+package com.alom.dorundorunbe.domain.achievement.dto.reward;
+
+public record RewardAchievementRequestDto(
+        Long userId,
+        Long achievementId
+) {
+    public static RewardAchievementRequestDto of(Long userId, Long achievementId) {
+        return new RewardAchievementRequestDto(userId, achievementId);
+    }
+}
