+++ src/main/java/com/alom/dorundorunbe/domain/achievement/dto/update/UpdateAchievementRequestDto.java
@@ -1,0 +1,11 @@
+package com.alom.dorundorunbe.domain.achievement.dto.update;
+
+public record UpdateAchievementRequestDto(
+        String name,        // 수정할 업적 이름 (Optional)
+        String rewardValue  // 수정할 보상 값 (Optional)
+
+) {
+    public static UpdateAchievementRequestDto of(String name, String rewardValue) {
+        return new UpdateAchievementRequestDto(name, rewardValue);
+    }
+}
