+++ src/main/java/com/alom/dorundorunbe/domain/achievement/exception/AchievementAlreadyExistsException.java
@@ -1,0 +1,8 @@
+package com.alom.dorundorunbe.domain.achievement.exception;
+
+
+public class AchievementAlreadyExistsException extends RuntimeException {
+    public AchievementAlreadyExistsException(String message) {
+        super(message);
+    }
+}
