+++ src/main/java/com/alom/dorundorunbe/domain/achievement/exception/AchievementConditionNotMetException.java
@@ -1,0 +1,7 @@
+package com.alom.dorundorunbe.domain.achievement.exception;
+
+public class AchievementConditionNotMetException extends RuntimeException {
+    public AchievementConditionNotMetException(String message) {
+        super(message);
+    }
+}
