+++ src/main/java/com/alom/dorundorunbe/domain/achievement/exception/AchievementNotFoundException.java
@@ -1,0 +1,7 @@
+package com.alom.dorundorunbe.domain.achievement.exception;
+
+public class AchievementNotFoundException extends RuntimeException {
+    public AchievementNotFoundException(String message) {
+        super(message);
+    }
+}
