+++ src/main/java/com/alom/dorundorunbe/domain/achievement/exception/RewardAlreadyClaimedException.java
@@ -1,0 +1,7 @@
+package com.alom.dorundorunbe.domain.achievement.exception;
+
+public class RewardAlreadyClaimedException extends RuntimeException {
+    public RewardAlreadyClaimedException(String message) {
+        super(message);
+    }
+}
