+++ src/main/java/com/alom/dorundorunbe/domain/achievement/exception/UserAchievementAlreadyClaimedException.java
@@ -1,0 +1,7 @@
+package com.alom.dorundorunbe.domain.achievement.exception;
+
+public class UserAchievementAlreadyClaimedException extends RuntimeException {
+    public UserAchievementAlreadyClaimedException(String message) {
+        super(message);
+    }
+}
