+++ src/main/java/com/alom/dorundorunbe/domain/achievement/exception/UserAchievementNotFoundException.java
@@ -1,0 +1,7 @@
+package com.alom.dorundorunbe.domain.achievement.exception;
+
+public class UserAchievementNotFoundException extends RuntimeException {
+    public UserAchievementNotFoundException(String message) {
+        super(message);
+    }
+}
