+++ src/main/java/com/alom/dorundorunbe/domain/achievement/repository/AchievementRepository.java
@@ -1,0 +1,12 @@
+package com.alom.dorundorunbe.domain.achievement.repository;
+
+import com.alom.dorundorunbe.domain.achievement.domain.Achievement;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface AchievementRepository extends JpaRepository<Achievement, Long> {
+
+    Optional<Achievement> findByNameAndCondition(String name, String condition);
+
+}
