+++ src/main/java/com/alom/dorundorunbe/domain/user/domain/User.java
@@ -1,13 +1,17 @@
+
 package com.alom.dorundorunbe.domain.user.domain;
 
+import com.alom.dorundorunbe.domain.ranking.domain.Ranking;
 import com.alom.dorundorunbe.global.util.BaseEntity;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.Getter;
+import lombok.Builder;
 
 @Entity @Getter
+@Setter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
@@ -16,6 +20,7 @@ public class User extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "user_id")
     private Long id;
 
     // 임의로 추가한 필드 -> kakao 로그인 구현 시 수정
@@ -25,11 +30,38 @@ public class User extends BaseEntity {
     @Column(nullable = false, unique = true, length = 32)
     private String nickname;
 
+    @Column(nullable = false, length = 32)
+    private String name;
+
+    @Column(nullable = false, length = 50)
+    private String email;
+
     private int age;
+
+    private String password;
+
+    @OneToOne(mappedBy = "user") //ranking 객체를 참조하는데, 이때 ranking 내의 user가 관계 주인
+    private Ranking ranking;
 
     @Enumerated(EnumType.STRING)
     private Gender gender;
 
     @Column(nullable = false)
-    private int cash;
+    private Long cash;
+    @Column(nullable = false,length = 30)
+    private String background;
+
+    public void updateBackground(String newBackground) {
+        this.background = newBackground;
+    }
+
+    public void addCash(int rewardValue){
+        cash += rewardValue;
+    }
+
+    public void updateCash(Long cash) {
+        this.cash = cash;
+    }
 }
+
+
